requirements.txt
flask==2.3.3
flask-cors==4.0.0
feedparser==6.0.10
requests==2.31.0
openai==1.3.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
schedule==1.2.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
nltk==3.8.1
textblob==0.17.1
newspaper3k==0.2.8
pytz==2023.3
sqlalchemy==2.0.23
flask-sqlalchemy==3.1.1
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0

